fix: gitpod enviroment
The prebuilt gitpod environment was missing opensafely, because
apparently only /workspace is preserved from a prebuilt image, and
opensafely was being install in $HOME.

So, we change this to install opensafely into a virtualenv in
/workspace, which should be preserved, and set the PATH to use it by
default.

I also switched to using the `gp env` tool to set MAX_WORKERS.

I did also try custom Dockerfile, but it didn't work as expected.